Rough prototype for solving n-car optimization
Generalized algorithm to work for n cars. Added an iterative modification step to ensure the Hessian is PSD, but had to take it out since finding the eigenvalues is slow for problems with multiple cars. A constant regularization will need to be added since we do have negative eignevalues sometimes. Convergence is very poor. Should be able to plot and animate up to 5 cars/trajectories now. Next step is to improve convergence, then add soft-constraint on collisions.