Code Restructuring Pt.1
Replace CarObjective.m with a class that acts as a controller for n cars. This controller is now queried for dynamics updates and cost information. Trajectories are now held for each individual car in a Car object. The Car object is now responsible for maintaining trajectory and control info, providing dynamics updates given a control at a timestep, calculating the trajectory and control costs, and for drawing. The controller combines individual car costs and combines it with a collision cost (which has been disabled temporarily) to give a global cost. A lot of the code has been simpplified as a result.