Workaround for dynamics breakdown and usability improvements + code cleanup
The issue where the dynamics step breaks only happens when we are trying a new control sequence u + alpha*p = u' where u' has some radically large values that break assumption in our dynamics model. As a workaround, if we assign a cost of inifnity when this happens (instead of erroring out), we simply try again with alpha/2 until u' is valid. A key to this is that dynamics never break down when we are calling it through finite_differences since u* = u + lambda will never be radically different from the initially valid u. This change also moves teh trajectory legality check to the controller, and it commonizes all references to angle or acceleration limits. We now have two plots, one for each optimization method, whereas before one would overwrite the other. I have also been able to improve convergence by decreasing the weight of the control soft-constraints, which wasn't possible before due to the dynamics breakdown.