Gui Initial Commit
Toggle Buttons 'Initialize' and 'Randomize' - the former starts the optimization with pre-defined controls, the latter randomizes controls. Click on the buttons again to stop the optimization. Three modes of operation, Newton, DDP, or Newton and DDP (default). Added the ability to run single-step increments of Newton and DDP optimization so that we can see each method's progress side-by-side in the same figure concurrently rather than in sequence. Callbacks are also used to display iteration number and costs. 'Time step' and 'Check something' does nothing.